Cleanup structure and add JSDocs to all public APIs
