+++ packages/react-tinacms-contentful/src/authentication/modals/ContentfulAuthModal.tsx
