+++ packages/react-tinacms-contentful/src/hooks/useContentfulManagement.ts
@@ -1,6 +1,12 @@
 import type { ClientAPI } from 'contentful-management/dist/typings/create-contentful-api';
 import { useContentful } from './useContentful';
 
+/**
+ * Retrieves the Contentful Management SDK from the CMS
+ * 
+ * @param spaceId When using multiple spaces, specify the space you need the client for (Optional)
+ * @returns 
+ */
 export function useContentfulManagement(
   spaceId?: string
 ): ClientAPI | undefined {
