+++ packages/react-tinacms-contentful/src/index.ts
@@ -1,5 +1,5 @@
 export * from 'tinacms-contentful';
-export * from './components';
+export * from './authentication';
 export * from './hooks';
 export * from "./media";
 export * from './providers/TinacmsContentfulProvider';
