+++ packages/react-tinacms-contentful/src/media/media-store.ts
@@ -1,35 +1,15 @@
 import { ContentfulClient, ContentfulMediaStore as BaseContentfulMediaStore, ContentfulMediaStoreOptions } from "tinacms-contentful";
-import { MediaArchiveAction } from "./actions/archive";
-import { MediaFilterAction } from "./actions/media-filter";
 
 export interface ContentfulReactMediaStoreOptions extends ContentfulMediaStoreOptions {
   filter?: boolean;
   archive?: boolean;
 }
 
+/**
+ * @deprecated Please us ethe ContentfulMediaStore class instead
+ */
 export class ContentfulReactMediaStore extends BaseContentfulMediaStore {
   constructor(ContentfulClient: ContentfulClient, options: ContentfulReactMediaStoreOptions) {
     super(ContentfulClient, options);
-
-    this.actions = [
-      ...this.actions,
-    ]
-
-    if (options.filter !== false) {
-      this.actions.push(MediaFilterAction)
-    }
-
-    if (options.archive !== false) {
-      this.actions.push(MediaArchiveAction)
-    }
-
-    if (options?.actions && options?.actions?.length > 0) {
-      this.actions = [
-        ...this.actions,
-        ...options.actions
-      ]
-    }
   }
-
-  actions: any[] = [];  
 }
