+++ packages/tinacms-contentful/src/ContentfulRestApi/management/operationsGraph.ts
@@ -1,6 +1,7 @@
 import { v4 } from 'uuid';
 import { ContentType, ContentTypeLink, Sys } from "contentful";
 import { getFieldsWithReferences } from "./locale";
+import { findReferenceKeys, hasChanged, hasNode } from './utils';
 
 export interface OperationsGraph {
   nodes: Operation[];
@@ -35,30 +36,14 @@ export type GraphOptions = {
   contentType?: ContentType
 }
 
-const isEntry = (entry: Entry<any>) => typeof entry.sys !== "undefined" && entry.sys.type === "Entry"
-const isEntries = (entries: Entry<any>[]) => entries.findIndex && entries.findIndex(entry => isEntry(entry)) > -1
-const hasNode = (operations: OperationsGraph, node: Entry<any>) => operations.nodes.findIndex(currentNode => currentNode.sys.id === node.sys?.id) > -1 || false
-
-function hasChanged(initial: Entry<any> | null, updated: Entry<any> | null) {
-  const a = JSON.stringify(initial?.fields) ?? false
-  const b = JSON.stringify(updated?.fields) ?? true
-
-  return a !== b;
-}
-
-const findReferenceKeys = (entry: Entry<any>, contentType?: ContentType) => {
-  if (!entry) return [];
-
-  return Object.keys(entry.fields)
-    .filter(fieldName => {
-      const field = entry.fields[fieldName]
-      const fieldDefinition = contentType && contentType.fields.find(field => field.id === fieldName);
-      const isReference = fieldDefinition?.type === "Link" || Array.isArray(field) ? isEntries(field) : isEntry(field)
-
-      return isReference
-    })
-}
-
+/**
+ * Compares initial state to updated state and decides if a create, update, or dereference operation
+ * 
+ * @param initial Initial state of entry
+ * @param updated Updated state of entry
+ * @param options Graph options
+ * @returns An operation or null if no operation is necessary
+ */
 const createContentfulOperation = (initial: Entry<any> | null, updated: Entry<any> | null, options?: GraphOptions): Operation | null => {
   const isDereference = initial !== null && updated === null;
   const isCreate = initial === null && updated !== null;
@@ -99,6 +84,16 @@ const createContentfulOperation = (initial: Entry<any> | null, updated: Entry<an
   return null
 }
 
+/**
+ * Creates a node in the graph if a valid operation is found
+ * 
+ * @param graph The graph to add the node to
+ * @param initial The initial state of the node (entry)
+ * @param updated The updated state of the node (entry)
+ * @param parent If the node (entry) has a parent node (entry)
+ * @param options Graph options
+ * @returns The updated graph
+ */
 const createNode = (graph: OperationsGraph, initial: Entry<any> | null, updated: Entry<any> | null, parent: Entry<any> | null = null, options?: GraphOptions) => {
   if (initial === null && updated === null) return;
 
@@ -111,8 +106,17 @@ const createNode = (graph: OperationsGraph, initial: Entry<any> | null, updated:
       createEdge(graph, parent.sys.id, operation.sys.id)
     }
   }
+
+  return graph;
 }
 
+/**
+ * Adds an "edge" to the graph, representing a relationship from a parent to a child
+ * @param graph The graph to add the edge to 
+ * @param from The parent node
+ * @param to The child node
+ * @returns The updated graph
+ */
 const createEdge = (graph: OperationsGraph, from: string, to: string) => {
   if (!Array.isArray(graph.edges[from])) graph.edges[from] = [];
