+++ packages/tinacms-contentful/src/ContentfulRestApi/pagination.ts
@@ -2,6 +2,15 @@ import {
   ContentfulCollection,
 } from 'contentful';
 
+/**
+ * Completely resolves a Contentful collection to an array of items
+ * 
+ * @param query The query to fetch the collection
+ * @param collection The current collection
+ * @param items The current array of items
+ * @param getNextPage The logic to fetch the next page
+ * @returns An array of items
+ */
 export async function handleCollection<CollectionEntity = unknown>(
   query: any, collection: ContentfulCollection<CollectionEntity>, items: CollectionEntity[], getNextPage: (query: any) => Promise<ContentfulCollection<CollectionEntity>>
 ): Promise<[CollectionEntity[], ContentfulCollection<CollectionEntity>]> {
